Feat: refactored more code
