+++ packages/ledger/branchdag/utils.go
@@ -1,0 +1,11 @@
+package branchdag
+
+type utils struct {
+	branchDAG *BranchDAG
+}
+
+func newUtil(branchDAG *BranchDAG) (new *utils) {
+	return &utils{
+		branchDAG: branchDAG,
+	}
+}
