+++ packages/ledger/storage.go
@@ -211,7 +211,7 @@ func (s *storage) CachedConsumers(outputID utxo.OutputID) (cachedConsumers objec
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-// region DB prefixes //////////////////////////////////////////////////////////////////////////////////////////////////
+// region db prefixes //////////////////////////////////////////////////////////////////////////////////////////////////
 
 const (
 	// PrefixTransactionStorage defines the storage prefix for the Transaction object storage.
